kernel-abi-check: fix when meta gets updated, but not kernel
This is a rather gross hack. Try to handle the case where the meta
package for a kernel needs to be spun immediately after a release, but
the main kernel does not (this happened for focal). Basically, compare
meta- and signed- ABI versions between with the versions in the their
respective pocket *or* with the version in the initial release pocket.
Handling the reporting is a little wonky.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>