reverted-CVEs list: make sync_from_usns stop whinging about CVE-2020-0543
We referenced CVE-2020-0543 in these kernel USNs because they were
modified to support the boot option to disable the mitigation for
CVE-2020-0543 and to report the status of the vulnerability. But the
kernel wasn't really "affected" by the issue. Mark it as reverted since
we don't really have an appropriate exception list for this situation.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>