scripts/check-cves: Add support for doing CVE mistriage retriage
This adds a new command-line option --mistriage which specifies the number
of mistriaged CVEs to process. Like scripts/report-mistriaged-cves.py, a
mistriaged CVE is one where in Ubuntu the CVE exists in
ignored/not-for-us.txt but where Debian have triaged the CVE against a
particular package *and* where at least one of those packages exists in a
supported Ubuntu release.

When a CVE is 'add'ed during mistriage, check-cves will automatically try
and remove it from ignored/not-for-us.txt via sed.

When a CVE is 'ignore'd during mistriage, check-cves will add an entry to
ignored/ignore-mistriage.txt so that subsequent --mistriage runs will not
keep resurfacing the CVE for mistriage detection.

Finally, CVEs are chosen for mistriage detection so that newest CVEs are
more likely to be selected over older ones so that we can process the most
relevant ones first.