generate-oval: fix logic around lines to ignore w/out alpha
The addition of the alpha option broke the logic for lines to
ignore when parsing OVAL output by basically not ignoring lines that
should be ignored when the --alpha option has not been passed on the
command line. Fix this to drop the line if config.alpha is not set or
else use the addition "/esm" logic if config.alpha is set.

But I'm not entirely sure of the alpha logic is supposed to be doing, so
this may be wrong.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>