kernel/CVE-2020-14331: set priority to negligible
As this requires a config option that no kernel has turned on, I decided to
lower its Priority. I did it instead of just marking them as not-affected
because we don't have good tooling to detect security regressions due to
config options changes. So, we'd rather just backport the fix with a very
low priority.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>