generate-oval: fix bug affecing USNs with multiple binaries
Report USNs with multiple binaries that have different version of
which only a subset are vulnerable as vulnerable. The <criteria>
section was requiring all sub-criterion to report true (have at least
one package installed with a version less than the version in the
USN) to report the system as having a package with a pending fix;
instead return True if any of the sub-criterion return true.

[sbeattie: expanded commit message a bit.]

Acked-by: Steve Beattie <steve.beattie@canonical.com>