generate-oval: use atomic rename to avoid cache corruption
The host the generate-oval script runs on is over-subscribed, so
multiple instances of it can be running at the same time. When writing
the package cache, write to a temporary file and rename it into place to
avoid multiple instances of the script trying to write out at the same
time.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>