kernel/CVE-2020-28374: adjust priority for likely non-backing store kernels
Kernels in clouds and laptop or low power devices are less likely to be
multi-backstore iscsi or other LIO backends, so drop the priority on
addressing this issue in those kernels to medium.

It is planned that this issue will be addressed in the current kernel
SRU cycle, resulting in kernels published on 2021-01-25.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>