kernel/CVE-2020-29373: use io_uring introducing commit as break
Any operations that handle paths are vulnerable. sendmsg was the earliest one
that was introduced. But in case we want to backport any of the other
operations without backporting sendmsg, we risk missing vulnerable kernels.

Using the commit that introduced io_uring is safer, in this case. There should
be no supported kernel with io_uring but no vulnerable operation right now,
anyway.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>